The great re-working, part I.
This comes out of the realization that the only files that can affect
the generated INDEX are the per-port and per-category Makefiles, any
other makefiles that they include (which are shown via the
.MAKEFILE_LIST variable) plus the pkg-descr file for each port.

The pkg-descr files are only needed for the WWW setting in most of
them.  This could be moved into the 1ary Makefile as a simple variable
setting which would save some time consuming file-IO.

Tracking which Makefiles and pkg-descr files are included by what
ports has necessitated creating two more object types to be held in
the cache.

Conversely, this change means it is no longer necessary to track
master-slave relationships (implicit in the way Makefiles are
included), and it should make tracking options changes etc. more
automatic.

Maintaining lists of this-uses-that efficiently in those ports
necessitates creating a small ListVal class, for manipulating randomly
access lists of unique values.

Work in progress.  This marks getting cache-init to run to completion
under the new regime.