Expand the 'stuff' item into individual components -- this is preparatory to adding a 'make-readmes' command.
Add a COMMENT value to Category objects (for make_readmes)

Drop the {ACCUMULATED} hash from the Tree object -- it was never used,
and on reflection, it's going to be more efficient to just update the
objects in {LIVE} directly.

Use $tree->allports_data() when printing out the INDEX or SHLIB
indexes.