Drop OPTIONS from the list of variables we extract the values of: the ports tree is fully OPTIONSng-ified and has been for some time, so this variable is always empty now.  Use OPTIONS_DEFINE in its place.
Add OPTIONS_FILE to the list of variables.  This is now the official
location for the file where option settings are stored, and will now
be used for any port where OPTIONS settings are modified.  However the
ports still respects the settings from the old OPTIONSFILE location,
so process both of those locations.